[_toplevel.BaseTypes] Move parseInt and parseFloat here
